Fixed a bug in binarySearch.
Fixed codePointClass function in unicodeIntervals.p. The code was making
an incorrect bounds check on an interval. Certain identifiers were
correctly recognized, while others were not.

Implemented proper error behavior on any Unicode code point that is not
white space or a letter (although the letter processing is not yet
implemented). Digit processing is not handled correctly either.