Clean up parsing of doclets.
Essentially, the rule is that you can begin any line of a doclet with arbitrary amounts of white space plus one asterisk. All the white space and the asterisk are removed from the doclet before parsing any other aspect of it. However, the parser was not enforcing the start-of-line rules when inside a {@...} sequence.