Fix infinite loop in register allocation.
The logic for sorting out which temporaries are in which registers when sub-expressions don't put values in quite the right registers was wrong. In trying to keep a count of how many temporaries are right, a 'doesFit' count is maintained, but in rare cases 'doesFit' could grow larger than the total number of temporaries, which is incorrect.