Fix a bug where automatic objects that had a type with an unimplemented abstract function were being allowwed.
Expanded date formatting options to include E and x formats (for day of week and time zone formatting).

Added functions to the http client support that format rfc822 dates.
Made the userAgent string sent from the HttpClient class user-settable.
Added support for injecting query parameters and headers such that they can override one another and are properly escaped and formatted.
Made the set of headers automatically injected in HttpClient merely default values.
Refactored the URI parsing logic into the Uri class and modified HttpClient to use it.
Extended the Reader class to support length and reset semantics (but not necessarily on all Readers).
Allow HttpClient post operations to use a Reader for the body, thus eliminating the need to read a disk file entirely into memory.
Modified the weekDay field of Date to be an enum.