Change the http.Uri class to better comport with RFC3986.
There is now a strict and a relaxed option. The strict option enforces RFC3986 compatibility, while relaxed mode allows for URI formats that are closer to what a browser will accept.

Fix a bug with object aggregates. They now allow you to include inherited public members in the initialized fields of the aggregate.