Fix a bug with deleteAll on maps leading to a double-free.
Added 'homeDirectory' to the storage namespace.

Added an ini-file parser.
Fixed a bug with handling of empty path strings in Uri parsing.
Made the default host header hide the port if it is the default port for the scheme being used.
Made FileReader's rewindable.
Fixed a bug in SHA_256 calculation is the Init/Update/Final combination.
Added handling for SIGTRAP + tkill exceptions.