Make setDefaultLocale atomic.
The setDefaultLocale did have a race condition if two threads are trying to set the default locale at the same time. It was possible for both threads to get the defaultLocale that was defined before either call was made. Now, one of the threads will get the default locale set by the other.