Fix code generation bug with stack synchronization.
After stack and register arguments are separated, the amount of stack space needed is calculated. The exact amount was not critical, only the amount modulo 16. The calculation miscalculated the stack alignment because it counted an ellipsis argument list both for its number of elements to push and the size of the argument as if it had not been an ellipsis argument. There needed to be an 'else'. 

The problem stems from the fact that the code generator is trying to avoid having to maintain machine instructions in an editable form after code generation. If it were relatively easy to insert instructions, then one could use a simple scheme of generating the actual instructions in a code sequence, then insert the stack adjustment instructions where they need to go after you know what you actually did push onto the stack.
To avoid that work, the code generator must first count the instructions that will be generated and track their stack effects, then generate the actual instructions later on. 
Of course, instruction scheduling will require exactly a need to generate instructions in a way that later insertions would be easy. So, we can revisit this error-prone approach later.