Create a Lock node type for 'lock' statements.
Simplified the structure of the parse tree. Instead of overloading Block objects, with fairly clumsy manipulations needed, the Lock object more clearly encodes the element of the statement and also makes adding the Monitor.release call to be added to the exception handling appropriately.

The exception handling additions are not in the checkin. This is just the refactoring.