Fixed a codegen bug with complex expressions.
The code paths are complex enough that conditional expressions evaluated with a compare method can generate a corrupted push instruction that magically turns into a Rex prefix byte that happens to work for the next instruction. If a large class compare method is involved, the code sequence to allocated registers was mis-coded. By chance, the test code happened to work, but failed in other, important cases. The revised code properly attributes the temporary registers and the push and pop code paths for spills now handles the substring and substring16 classes (they are a special case in a switch that was failing without them).