Fix the pxi section type. The compiler was generating and loading the Windows section type even though it was compiled for the linux target.
Fix a bug in the thread initialization that was not setting the Parasol Thread parameter correctly for the main thread.

Improve the logic for parsing locale names to better fit the documentation.
Fix a bug where byte.isLower() reported true for capital C and false for lower-case c.
Improve code generation on stack frame initialization and on constructor object initialization (avoids calling memset on small amounts of memory).
Modified dumppxi to allow for dis-assembly of a pxi (though it is missing symbols).