Fix a bug in the LeaksHeap that was not removing empty Sections.
Change the main thread init logic to run under the normal heap even when the LeaksHepa is requested, because harvesting the main thread object seems to cause problems.

Fix a memory leak involving setting a temp as a live symbol when constructing an object from an object aggregate.
Started to add some infrastructure to identify objects that can't be copied, but there is no way to easily or obviously mark a class as not copyable.
Fixed up logic for returning strings and vectors by value so that if they are simple temps or local variables about to be destroyed, avoid that and copy their value cheaply.
Consolidate the type methods 'conforms' and 'namesSameObject' as they calculate the same thing: whether two Node trees describe the same lvalue expression.
Cleaned up some of the displays in exception stack traces.
Fixed a leak in vector templates where certain code paths are not destroying live objects.
Improved the display of symbol names so that for methods or members of a template, the correct instantiation arguments are listed in the disassembly.
Fixed a leak in for-in loops where the key was not getting initialized correctly (leading to a leak).
Removed obsolete 'built-ins' section of disassembly. Also stopped printing the list of relocations as they are not interesting and are quite long.