Made the compiler main delete the ParasolCommand objet it allocates.
When using the LeakHeap, the stdin, stdout and stderr streams are re-opened to the console, which allows for cleaner error handling at that stage.

Improved dumps of the LeakHeap when CorruptHeap conditions are found.
Fixed a memory leak resulting from string additions.
Fixed several leaks associated with the compilation process (primarily due to allocations not using one of the NoRelease pool allocators.
Added checks for valid annotations and fixed a bunch of misspelled or misplaced annotations.
Fixed a couple of memory leaks in maps associated with not destroying keys in some code paths.
Fixed a codegen bug where flags objects were writing into memory in the compiled static data area (during compile time) where they shouldn't.