Fixed switch statements and destructors associated with initialized symbols.
Fixed several memory leaks.

Fixed a bug in the allocation of CodeSegment's for switch-cases. The net effect was to let the memory of the case labels vector get misallocated and corrupt the heap.
Marked break, continue and throw statements as not falling through, so that scope can be properly terminated around them.