Get HTTP-based RPC's functioning.
End-to-end operation of RPC's is now working. The proxy class mechanism needs to be generalized (the call-out to RPC transport layer needs to be turned into an interface or base class). The range of types that have marshallers needs to be expanded. Support for short-circuit response to void methods needs to be implemented. An appropriate security mechanism needs to be added.

Implied destructors are now added for classes that implement an interface (so that the interface can be used to delete the object). This is actually in the category of how did this ever work? It's possible that all of the classes that implement interfaces so far have destructors for other reasons and it was only in testing RPC's that this surfaced.
Two substring constructors had buggy implementations.