Relax the rule that the RHS of an assignment of an interface must be an lvalue.
This rule was introduced, probably because the interface cannot be reliably converted to it's underlying class, and the reutrn value of a function may not have any other live references. However, you CAN delete the object an interface pointer points to, so the interface object is a valid reference.

Also adjust the rpc API to make it easier to initialize the Web service server-side objects.