Fix rpc_test, refine RPC interfaces and fix bug with constructor calls
The last commit included a failing rpc_test. This triggered a revamp of the RPC initialization and shutdown APIs, which in turn triggered another bug.

The other bug was caused because an obkect initialized with a constructor (empty parens) call for a class ith no constructors. The fix effectively makes such an initializer a no-op. The bug was that an interface for the object was not getting initialized. Now it is.