Add a debug namespace. Fix two code gen bugs.
First code gen bug has been in place for a long time and involved the creation of method maps for classes. This determines, for example, the content of vtables. Sometimes, the algorithm would think a class had built its method maps, but it had only started to. The fix involved adding a 'buildfing' state.
Second code gen bug consisted of involved another case where method maps were not getting created in time and resulted in a missing vtable, but to certain lazy evaluations that left an open code path.