1. Changed the compile rulles. In version 0.14, the pc command would include source files in directories under the one containing the main script file.
This creates all sorts of problems with making simple script or test files.
Plenty of places in the file system might have subdirectories containing Parasol sources
unintentionally.

Undur the new rules, the pc command will only include the main script file and resolve any
references to namespaces in the current Parasol context (defined by the environment variable
PARASOL_CONTEXT).

However, new command-line options for including one or more directories containing Parasol
sources in a compile.

The -r/--recursive flags have been added to apply the old rules (although this flag will
probably be removed before long.

2. Also reorganized the source tree to be a little more aligned with the build structure.
Thus there is now a cmd directory continaing the main files of the paradoc, pbug, pbuild,
pcontext and pc commands.

There is also a new lib directory containing the sources for the build, debug and documentation
packages.

The changes to the source tree had to be reflected in the build files. Since the Parasol
documentation describes the various library facilities in terms of their namespaces,
little had to be done.

3. The compiler tests now do not offer a 'compile-from-source' option since it onlly slowed down
compiles, when a compiled version of pc was available from the build tree.

4. There is a bug in the parser/scanner somewhere that appears to corrupt memory and definitely
causes a segmentation violation (SIGSEGV) error in some Parasol builds. This update does
not fix this problem as it does not interfere with the official builds.

5. There is a bunch of work on the new stack trace logic. It is incomplete, but correctly identifies
the start of a thread stack (newer libc has  made the symbol it relied on private).

The logic does a better job of finding and reporting the correct stack frames, as long as the
stack is not corrutped (as this build bug seems to produce).

6. Fixed a scanner bug in semi-colon elision. A line that ends with a line comment (//) did not
set the correct flag, so no semi-colon was generated when one should be.

7. Fixed a bug in pbuild where single-threaded runs would not handle dependencies between packages
and would produce a hang in the build. The solution was to do a tpoplogical sort to feed
the list of products to the worker thread pool in the order that, if run by a single thread,
would correctly build the packages and applications in the correct order.

Fixed another issue with pbuild where it would not force a build of makefiles on an official
build, when all products should be recompiled with correct version strings.

8. Fixed a pbug command-line processing bug involving the application arguments being passed.
If the application wants arguments beginning with a dash character, pbug would not recognize
the difference between pbug options and target program options.

9. The pbug memory map now returns the memory segment containing the symbol when one is looked up.
This is used in pbug to tell when it is being run against older libc/libpthread combinations.

10. Added logic to pbug to parse a debug script and validate it's contents.